requirements.txt
Django==1.10.1
django-registration==2.2
whitenoise==3.3.0
readability-lxml==0.6.2
beautifulsoup4==4.4.1
django-registration-defaults==0.4.4
